+++ openarcrt/openaccrt.cpp
@@ -776,7 +776,7 @@ HI_error_t HI_register_kernel_arg(std::string kernel_name, int arg_index, size_t
 	return return_status;
 }
 
-HI_error_t HI_kernel_call(std::string kernel_name, int gridSize[3], int blockSize[3], int async, int num_waits, int *waits) {
+HI_error_t HI_kernel_call(std::string kernel_name, size_t gridSize[3], size_t blockSize[3], int async, int num_waits, int *waits) {
 	HI_error_t return_status;
 	const char *kernelName = kernel_name.c_str();
 #ifdef _OPENARC_PROFILE_
